refactor(core): extract scanner pure helpers + upserts
Final slice of step 6. Pulls the genuinely portable half of
`commands/scan.rs` (~1100 of its 2045 LOC) into a new
`waveflow_core::scanner` module:

- `scanner::extract` — `ExtractedFile`, `ExtractedCover`, the
  tag-to-struct mapping (`extract_cover`, `extract_folder_cover`,
  `extract_artist_image`, `extract_rating`, `extract_musical_key`,
  `extract_album_artist`, `extract_compilation_flag`),
  `file_type_label`, `extension_for_mime`, `hash_file`, and the
  `AUDIO_EXTENSIONS` whitelist. The cover-pipeline helpers now point
  at `crate::artwork::thumbnails::spawn_thumbnail_job` (in-crate)
  instead of `crate::thumbnails`.
- `scanner::upserts` — `canonical_name`, `split_artist_name`,
  `now_millis`, `VARIOUS_ARTISTS_LABEL`, the five `upsert_*`
  helpers (artwork / artist / album / genre / artist_list),
  `resolve_album_artist` (the Album-Artist grouping policy),
  `link_local_artist_image`, `maybe_link_artist_images`, and the
  `merge_implicit_compilations` post-scan promotion pass.

All eight scanner unit tests (folder-cover priority, JPEG
normalisation, artist-image walk-up + canonical-name stem match,
hash-addressed write) move alongside the code they cover into
`core/src/scanner/extract.rs`.

What stays in `crates/app/src/commands/scan.rs`:

- `ScanProgress`, `ScanSummary`, `maybe_emit_progress` — the
  `scan:progress` Tauri-event shape and the throttled emitter.
- `extract_dsd_file` + the `extract_file` dispatcher — both still
  call into `crate::audio::dsd::*`, which migrates to core in
  step 7. The DSD-aware extractor stays app-side until then.
- `scan_folder` (`#[tauri::command]`), `scan_folder_inner` (the
  walkdir + `buffered`-stream orchestrator), `rescan_local_artist_images`,
  `ArtistImageScanSummary`.

App-side wiring:

- A `pub use waveflow_core::scanner::canonical_name;` re-export at
  the top of `commands/scan.rs` keeps the two cross-file imports
  in `commands/{radio,similar}.rs` resolving (`crate::commands::scan::canonical_name`).
- `commands/edit.rs` updates its scanner import to
  `waveflow_core::scanner::*` directly (no re-export shim — the
  call site is the only one outside scan.rs that pulled on
  multiple helpers).

`crates/core/Cargo.toml` picks up `lofty = "0.24"` (tag parser).

Scanner orchestration (`scan_folder_inner` decoupling from
`tauri::AppHandle` via a `ScannerEventSink` trait) is the obvious
follow-up but stays outside the Phase 1.a brief — `scan_folder_inner`
is fine app-side for now.

`cargo check --workspace --all-targets` clean; `cargo test
--workspace` 111 passed / 0 failed (83 app + 28 core).

Refs #128.
