+++ src-tauri/crates/app/src/commands/scan.rs
@@ -1,24 +1,34 @@
 use std::collections::HashMap;
 use std::fs;
-use std::io::Read;
 use std::path::{Path, PathBuf};
 use std::time::UNIX_EPOCH;
 
-use chrono::Utc;
 use futures::StreamExt;
-use lofty::file::{FileType, TaggedFileExt};
-use lofty::picture::MimeType;
+use lofty::file::TaggedFileExt;
 use lofty::prelude::{Accessor, AudioFile};
-use lofty::tag::{ItemKey, Tag, TagType};
 use serde::Serialize;
 use sqlx::SqlitePool;
 use tauri::Emitter;
 use walkdir::WalkDir;
 
+use waveflow_core::scanner::{
+    extract_album_artist, extract_compilation_flag, extract_cover, extract_folder_cover,
+    extract_artist_image, extract_musical_key, extract_rating, file_type_label, hash_file,
+    link_local_artist_image, maybe_link_artist_images, merge_implicit_compilations, now_millis,
+    split_artist_name, upsert_album, upsert_artist, upsert_artist_list, upsert_artwork,
+    upsert_genre, ExtractedFile, AUDIO_EXTENSIONS, VARIOUS_ARTISTS_LABEL,
+};
+
 use crate::{
     error::{AppError, AppResult},
     state::AppState,
 };
+
+// Re-export `canonical_name` so existing call sites
+// (`crate::commands::scan::canonical_name` in `commands/radio.rs` and
+// `commands/similar.rs`) keep resolving after the helper moved to
+// `waveflow_core::scanner` in step 6.d.
+pub use waveflow_core::scanner::canonical_name;
 
 /// Payload of the `scan:progress` Tauri event. Emitted by
 /// `scan_folder_inner` every ~25 processed files (and once at the end)
@@ -67,19 +77,6 @@ fn maybe_emit_progress(
     );
 }
 
-/// Extensions considered "audio files" by the scanner. Limited to
-/// formats the symphonia + cpal engine can actually decode and play,
-/// so the library never displays tracks that would error at play
-/// time. Opus / WMA / AIFF are intentionally absent — symphonia
-/// doesn't ship a mainline decoder for Opus, WMA is Microsoft
-/// proprietary, and AIFF isn't in the default feature set.
-const AUDIO_EXTENSIONS: &[&str] = &[
-    "mp3", "flac", "wav", "ogg", "oga", "m4a", "mp4", "aac",
-    // DSD: handled by the in-tree audio::dsd pipeline (symphonia
-    // doesn't decode DSD), with metadata read via audio::dsd::metadata.
-    "dsf", "dff",
-];
-
 /// Outcome of a `scan_folder` call, returned to the frontend so the UI can
 /// display a toast like "120 nouveaux titres · 3 mises à jour · 1 erreur".
 #[derive(Debug, Serialize, Default)]
@@ -95,437 +92,6 @@ pub struct ScanSummary {
     /// liked / playlist / play-event history) so the user can recover
     /// it by putting the file back.
     pub removed: u32,
-}
-
-/// Normalize a title/name for dedup purposes: lowercase, strip punctuation
-/// and collapse whitespace. Good enough to match "The Beatles" / "THE  BEATLES"
-/// or "the beatles!" onto a single canonical key without pulling in a proper
-/// Unicode normalization library.
-pub(crate) fn canonical_name(s: &str) -> String {
-    s.trim()
-        .chars()
-        .flat_map(|c| c.to_lowercase())
-        .map(|c| if c.is_alphanumeric() { c } else { ' ' })
-        .collect::<String>()
-        .split_whitespace()
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn now_millis() -> i64 {
-    Utc::now().timestamp_millis()
-}
-
-/// Stream the file through blake3 in 64 KiB chunks. Full-file hash — slower
-/// than a prefix hash but gives us reliable dedup across moved/renamed files.
-fn hash_file(path: &Path) -> std::io::Result<String> {
-    let mut hasher = blake3::Hasher::new();
-    let file = fs::File::open(path)?;
-    let mut reader = std::io::BufReader::new(file);
-    let mut buf = [0u8; 64 * 1024];
-    loop {
-        let n = reader.read(&mut buf)?;
-        if n == 0 {
-            break;
-        }
-        hasher.update(&buf[..n]);
-    }
-    Ok(hasher.finalize().to_hex().to_string())
-}
-
-/// Everything the scanner reads off disk for a single audio file. Populated
-/// inside `spawn_blocking` so the tokio reactor never stalls on I/O.
-struct ExtractedFile {
-    abs_path: String,
-    size: i64,
-    modified_ms: i64,
-    hash: String,
-    title: String,
-    artist: Option<String>,
-    album: Option<String>,
-    /// Raw Album Artist text from the source tag (`TPE2` / `aART` /
-    /// `ALBUMARTIST` / `Album Artist`). Used as the album-grouping
-    /// authority — when present, two tracks share an album even if
-    /// their per-track Artist tags differ (featurings, lead-vocal
-    /// rotations on K-pop EPs, etc.).
-    album_artist: Option<String>,
-    /// `TCMP` (ID3v2) / `cpil` (MP4) / `COMPILATION` (Vorbis / APE)
-    /// flag. When `true` the scanner uses a synthetic "Various
-    /// Artists" album artist so a true compilation merges its tracks
-    /// under a single album row even when no Album Artist tag exists.
-    is_compilation: bool,
-    genre: Option<String>,
-    year: Option<i64>,
-    track_number: Option<i64>,
-    disc_number: Option<i64>,
-    duration_ms: i64,
-    bitrate: Option<i64>,
-    sample_rate: Option<i64>,
-    channels: Option<i64>,
-    /// Bits per sample (16 for CD-quality, 24/32 for Hi-Res masters).
-    /// Lossy codecs (MP3, AAC) typically don't expose this — left as
-    /// `None` so the UI's Hi-Res badge logic can short-circuit without
-    /// inspecting the codec separately.
-    bit_depth: Option<i64>,
-    /// Short codec / container label inferred from the file type
-    /// (e.g. `"FLAC"`, `"MP3"`, `"AAC"`, `"WAV"`). Drives the format
-    /// chip on the player footer.
-    codec: Option<String>,
-    /// Tagged musical key when the file carries one (`TKEY` / ID3v2
-    /// or `INITIALKEY` / Vorbis-MP4-APE). Whatever notation the
-    /// upstream tagger chose stays as-is — could be `Am`, `F#`, or
-    /// the Camelot wheel `8A`.
-    musical_key: Option<String>,
-    /// Embedded cover art extracted and hash-addressed during the scan. Only
-    /// the first picture is kept (lofty exposes them in order and the first
-    /// is usually the `CoverFront`). `None` when the tag has no pictures.
-    cover_art: Option<ExtractedCover>,
-    /// Raw POPM byte (0-255) for ID3v2 files, or a normalised value
-    /// derived from the `RATING` text field for Vorbis/FLAC/MP4. `None`
-    /// when neither tag carries a rating.
-    rating: Option<u8>,
-}
-
-struct ExtractedCover {
-    /// Hex-encoded blake3 hash of the picture bytes — used as the filename
-    /// stem so identical artwork embedded in 20 tracks of an album yields a
-    /// single file on disk.
-    hash: String,
-    /// File extension matching the picture's MIME type (jpg/png/webp/...).
-    format: String,
-    /// Provenance label written to `artwork.source`. Either `"embedded"`
-    /// (lifted from the tag) or `"folder"` (sidecar cover.jpg / folder.png
-    /// / front.webp etc. next to the audio file).
-    source: &'static str,
-}
-
-/// Map lofty's `FileType` enum to a short uppercase label suitable
-/// for the UI's format chip. Falls back to `None` when lofty can't
-/// determine a recognized container — we'd rather hide the chip
-/// than print "Unknown".
-fn file_type_label(ft: FileType) -> Option<String> {
-    match ft {
-        FileType::Mpeg => Some("MP3".into()),
-        FileType::Flac => Some("FLAC".into()),
-        FileType::Mp4 => Some("AAC".into()),
-        FileType::Aac => Some("AAC".into()),
-        FileType::Wav => Some("WAV".into()),
-        FileType::Vorbis => Some("Vorbis".into()),
-        FileType::Opus => Some("Opus".into()),
-        FileType::Aiff => Some("AIFF".into()),
-        FileType::Speex => Some("Speex".into()),
-        FileType::Ape => Some("APE".into()),
-        FileType::WavPack => Some("WavPack".into()),
-        FileType::Custom(name) => Some(name.to_string()),
-        _ => None,
-    }
-}
-
-/// Pick a reasonable filename extension for lofty's MIME type enum. Unknown
-/// / exotic formats fall through to `"bin"` so the file is still written and
-/// the UI can decide what to do with it.
-fn extension_for_mime(mime: Option<&MimeType>) -> &'static str {
-    match mime {
-        Some(MimeType::Jpeg) => "jpg",
-        Some(MimeType::Png) => "png",
-        Some(MimeType::Gif) => "gif",
-        Some(MimeType::Bmp) => "bmp",
-        Some(MimeType::Tiff) => "tiff",
-        _ => "bin",
-    }
-}
-
-/// Extract the first picture from the given tag, hash-address it, and write
-/// it to `<artwork_dir>/<hash>.<ext>` if missing. Returns the identifying
-/// `ExtractedCover` or `None` when the tag has no pictures.
-///
-/// The write is idempotent: a file whose path already exists is assumed to
-/// match (because blake3 hashes are content-addressed), so we never
-/// overwrite on re-scan.
-fn extract_cover(tag: &Tag, artwork_dir: &Path) -> Option<ExtractedCover> {
-    let picture = tag.pictures().first()?;
-    let bytes = picture.data();
-    if bytes.is_empty() {
-        return None;
-    }
-    let hash = blake3::hash(bytes).to_hex().to_string();
-    let format = extension_for_mime(picture.mime_type()).to_string();
-    let out_path = artwork_dir.join(format!("{}.{}", &hash, &format));
-    if !out_path.exists() {
-        if let Err(err) = fs::write(&out_path, bytes) {
-            tracing::warn!(path = %out_path.display(), error = %err, "failed to write artwork");
-            return None;
-        }
-    }
-    crate::thumbnails::spawn_thumbnail_job(out_path, artwork_dir.to_path_buf(), hash.clone());
-    Some(ExtractedCover {
-        hash,
-        format,
-        source: "embedded",
-    })
-}
-
-/// Canonical filename stems searched for in the track's parent directory
-/// when the audio file carries no embedded picture. Order matters — the
-/// first match wins. Mirrors the convention used by foobar2000, MusicBee,
-/// Plex, Kodi, RustMusic.
-const FOLDER_COVER_STEMS: &[&str] = &["cover", "folder", "front", "albumart", "album", "artwork"];
-
-/// File extensions accepted as folder cover candidates. Limited to formats
-/// the `image` crate decodes via the features enabled in `Cargo.toml`, so
-/// every match downstream of this fn is guaranteed to be readable by the
-/// thumbnail pipeline.
-const FOLDER_COVER_EXTENSIONS: &[&str] = &["jpg", "jpeg", "png", "webp", "bmp", "gif", "tiff"];
-
-/// Look for a sidecar cover image (cover.jpg / folder.png / front.webp / ...)
-/// next to the track. Returns an `ExtractedCover` written to the shared
-/// artwork dir, hash-addressed like embedded pictures.
-///
-/// Used as a fallback when the audio file has no embedded picture — common
-/// for FLAC/WAV libraries ripped from CD where the artwork sits beside the
-/// tracks rather than inside them.
-fn extract_folder_cover(track_path: &Path, artwork_dir: &Path) -> Option<ExtractedCover> {
-    let parent = track_path.parent()?;
-    let entries = fs::read_dir(parent).ok()?;
-
-    // Index siblings by lowercased (stem, ext) for O(1) lookup against the
-    // priority lists above. Single read_dir pass — cheaper than 6×7 = 42
-    // `Path::exists` calls when the directory is large.
-    let mut candidates: std::collections::HashMap<(String, String), PathBuf> =
-        std::collections::HashMap::new();
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if !path.is_file() {
-            continue;
-        }
-        let stem = path
-            .file_stem()
-            .and_then(|s| s.to_str())
-            .map(|s| s.to_lowercase());
-        let ext = path
-            .extension()
-            .and_then(|s| s.to_str())
-            .map(|s| s.to_lowercase());
-        if let (Some(s), Some(e)) = (stem, ext) {
-            candidates.insert((s, e), path);
-        }
-    }
-
-    let picked = FOLDER_COVER_STEMS
-        .iter()
-        .flat_map(|stem| {
-            FOLDER_COVER_EXTENSIONS
-                .iter()
-                .map(move |ext| (stem.to_string(), ext.to_string()))
-        })
-        .find_map(|key| candidates.get(&key).cloned())?;
-
-    let bytes = fs::read(&picked).ok()?;
-    if bytes.is_empty() {
-        return None;
-    }
-    let hash = blake3::hash(&bytes).to_hex().to_string();
-    let format = picked
-        .extension()
-        .and_then(|s| s.to_str())
-        .map(|s| s.to_lowercase())
-        .unwrap_or_else(|| "jpg".to_string());
-    // Normalise `jpeg` to `jpg` so the artwork dir doesn't end up with two
-    // entries pointing at the same MIME.
-    let format = if format == "jpeg" {
-        "jpg".to_string()
-    } else {
-        format
-    };
-
-    let out_path = artwork_dir.join(format!("{}.{}", &hash, &format));
-    if !out_path.exists() {
-        if let Err(err) = fs::write(&out_path, &bytes) {
-            tracing::warn!(path = %out_path.display(), error = %err, "failed to write folder cover");
-            return None;
-        }
-    }
-    crate::thumbnails::spawn_thumbnail_job(out_path, artwork_dir.to_path_buf(), hash.clone());
-    Some(ExtractedCover {
-        hash,
-        format,
-        source: "folder",
-    })
-}
-
-/// Stems recognised as a sidecar artist photo at any ancestor level of a
-/// track. Matched verbatim (lowercased); a stem-aware match against the
-/// artist's canonical name handles the `<artist>.jpg` convention.
-const ARTIST_IMAGE_STEMS: &[&str] = &["artist", "performer", "band"];
-
-/// Maximum number of parent directories walked upward from the track to
-/// find an artist photo. Covers the two common layouts called out in
-/// issue #31:
-///   1. `<root>/<artist>/<album>/track.flac` → 2 levels up (`<artist>/`).
-///   2. `<root>/<album>/track.flac`         → 1 level up (`<album>/`),
-///      and even the album folder itself can hold an `<artist>.jpg`.
-///
-/// 3 covers the occasional `<root>/<artist>/<album>/CD1/track.flac` rip.
-const ARTIST_IMAGE_MAX_DEPTH: usize = 3;
-
-/// Look for a sidecar artist image next to the track. Walks up to
-/// `ARTIST_IMAGE_MAX_DEPTH` parent directories from `track_path` and
-/// accepts the first match where either:
-///   - the file stem is in [`ARTIST_IMAGE_STEMS`] (`artist.jpg`,
-///     `performer.png`, …), or
-///   - the file stem's canonical form equals `artist_canonical` (covers
-///     `Daft Punk.jpg` sitting at the root of a `Daft Punk/` folder).
-///
-/// Hash-addressed write into `artwork_dir` like every other cover so a
-/// later GC can dedup across artists and albums.
-fn extract_artist_image(
-    track_path: &Path,
-    artist_canonical: &str,
-    artwork_dir: &Path,
-) -> Option<ExtractedCover> {
-    if artist_canonical.is_empty() {
-        return None;
-    }
-
-    let mut current = track_path.parent();
-    for _ in 0..ARTIST_IMAGE_MAX_DEPTH {
-        let Some(dir) = current else { break };
-        if let Some(found) = find_artist_image_in_dir(dir, artist_canonical) {
-            return write_artist_image(&found, artwork_dir);
-        }
-        current = dir.parent();
-    }
-    None
-}
-
-fn find_artist_image_in_dir(dir: &Path, artist_canonical: &str) -> Option<PathBuf> {
-    let entries = fs::read_dir(dir).ok()?;
-    let mut named_match: Option<PathBuf> = None;
-    let mut stem_match: Option<(usize, PathBuf)> = None;
-
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if !path.is_file() {
-            continue;
-        }
-        let stem = path
-            .file_stem()
-            .and_then(|s| s.to_str())
-            .map(|s| s.to_lowercase());
-        let ext = path
-            .extension()
-            .and_then(|s| s.to_str())
-            .map(|s| s.to_lowercase());
-        let (Some(stem), Some(ext)) = (stem, ext) else {
-            continue;
-        };
-        if !FOLDER_COVER_EXTENSIONS.contains(&ext.as_str()) {
-            continue;
-        }
-        if canonical_name(&stem) == artist_canonical {
-            named_match.get_or_insert(path);
-            continue;
-        }
-        if let Some(rank) = ARTIST_IMAGE_STEMS.iter().position(|s| *s == stem) {
-            match &stem_match {
-                Some((current_rank, _)) if *current_rank <= rank => {}
-                _ => stem_match = Some((rank, path)),
-            }
-        }
-    }
-
-    named_match.or(stem_match.map(|(_, p)| p))
-}
-
-fn write_artist_image(picked: &Path, artwork_dir: &Path) -> Option<ExtractedCover> {
-    let bytes = fs::read(picked).ok()?;
-    if bytes.is_empty() {
-        return None;
-    }
-    let hash = blake3::hash(&bytes).to_hex().to_string();
-    let format = picked
-        .extension()
-        .and_then(|s| s.to_str())
-        .map(|s| s.to_lowercase())
-        .unwrap_or_else(|| "jpg".to_string());
-    let format = if format == "jpeg" {
-        "jpg".to_string()
-    } else {
-        format
-    };
-
-    let out_path = artwork_dir.join(format!("{}.{}", &hash, &format));
-    if !out_path.exists() {
-        if let Err(err) = fs::write(&out_path, &bytes) {
-            tracing::warn!(
-                path = %out_path.display(),
-                error = %err,
-                "failed to write artist image",
-            );
-            return None;
-        }
-    }
-    crate::thumbnails::spawn_thumbnail_job(out_path, artwork_dir.to_path_buf(), hash.clone());
-    Some(ExtractedCover {
-        hash,
-        format,
-        source: "folder",
-    })
-}
-
-/// Best-effort: link a freshly resolved local artist image to its `artist`
-/// row when the row has no artwork yet. Idempotent — re-running with a
-/// already-linked artist is a no-op (the `IS NULL` guard prevents
-/// overwriting a manually uploaded picture).
-async fn link_local_artist_image(
-    conn: &mut sqlx::SqliteConnection,
-    artist_id: i64,
-    cover: &ExtractedCover,
-) -> AppResult<()> {
-    let artwork_id = upsert_artwork(conn, &cover.hash, &cover.format, cover.source).await?;
-    sqlx::query("UPDATE artist SET artwork_id = ? WHERE id = ? AND artwork_id IS NULL")
-        .bind(artwork_id)
-        .bind(artist_id)
-        .execute(&mut *conn)
-        .await?;
-    Ok(())
-}
-
-/// Extract a 0-255 rating from a tag. POPM frames (ID3v2) are stored by
-/// lofty as raw `ItemValue::Binary` under `ItemKey::Popularimeter`: the
-/// frame body is `<email>\0<rating:u8><counter:u32+>`, so the rating is
-/// the byte right after the first NUL terminator. Vorbis/FLAC/MP4 expose
-/// `RATING` as plain text 0-100 which we rescale to 0-255.
-fn extract_rating(tag: &Tag) -> Option<u8> {
-    if matches!(tag.tag_type(), TagType::Id3v2) {
-        if let Some(bytes) = tag.get_binary(ItemKey::Popularimeter, false) {
-            let nul_pos = bytes.iter().position(|b| *b == 0)?;
-            return bytes.get(nul_pos + 1).copied();
-        }
-    }
-    if let Some(text) = tag.get_string(ItemKey::Popularimeter) {
-        let trimmed = text.trim();
-        if let Ok(val) = trimmed.parse::<u16>() {
-            let clamped = val.min(100);
-            return Some((clamped * 255 / 100) as u8);
-        }
-    }
-    None
-}
-
-/// Read the tagged musical key, if any. ID3v2 stores it as `TKEY`,
-/// Vorbis comments / MP4 / APE / WavPack as `INITIALKEY` — lofty
-/// unifies both behind `ItemKey::InitialKey`. Empty strings are
-/// coalesced to `None` so the UI's "—" placeholder kicks in
-/// instead of a blank cell.
-fn extract_musical_key(tag: &Tag) -> Option<String> {
-    let raw = tag.get_string(ItemKey::InitialKey)?;
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        None
-    } else {
-        Some(trimmed.to_string())
-    }
 }
 
 /// Build the standard `ExtractedFile` payload for a DSF / DFF file.
@@ -614,6 +180,9 @@ fn extract_dsd_file(
     })
 }
 
+/// Single-file extraction dispatcher: branches DSF/DFF onto the
+/// in-tree `audio::dsd` pipeline (symphonia/lofty don't read DSD) and
+/// everything else through lofty.
 fn extract_file(path: &Path, artwork_dir: &Path) -> Result<ExtractedFile, String> {
     let metadata = fs::metadata(path).map_err(|e| format!("metadata: {e}"))?;
     let size = metadata.len() as i64;
@@ -721,439 +290,6 @@ fn extract_file(path: &Path, artwork_dir: &Path) -> Result<ExtractedFile, String
         cover_art,
         rating,
     })
-}
-
-/// Pull the Album Artist tag and trim it. Lofty's `ItemKey::AlbumArtist`
-/// already abstracts the per-container mapping (`TPE2` / `aART` /
-/// `ALBUMARTIST` / `Album Artist`). Empty / whitespace-only strings are
-/// treated as missing so the grouping code falls back to the per-track
-/// Artist exactly like before.
-fn extract_album_artist(tag: &lofty::tag::Tag) -> Option<String> {
-    let raw = tag.get_string(ItemKey::AlbumArtist)?;
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        None
-    } else {
-        Some(trimmed.to_string())
-    }
-}
-
-/// Read the compilation flag (`TCMP` / `cpil` / `COMPILATION` / `Compilation`).
-/// Lofty stores the value as a stringified `0` / `1` regardless of the
-/// underlying container; anything that parses to a non-zero integer or the
-/// literal `true` is treated as "this is a compilation".
-fn extract_compilation_flag(tag: &lofty::tag::Tag) -> bool {
-    let Some(raw) = tag.get_string(ItemKey::FlagCompilation) else {
-        return false;
-    };
-    let trimmed = raw.trim();
-    if trimmed.eq_ignore_ascii_case("true") {
-        return true;
-    }
-    matches!(trimmed.parse::<i64>(), Ok(n) if n != 0)
-}
-
-/// Upsert an artwork row keyed on its content hash. Existing rows are
-/// returned as-is; new rows are inserted with the caller-supplied source
-/// label (`embedded`, `folder`, `deezer`, `user`...) so a future cleanup
-/// job can distinguish scanner-extracted art from remote/manual files.
-pub(crate) async fn upsert_artwork(
-    conn: &mut sqlx::SqliteConnection,
-    hash: &str,
-    format: &str,
-    source: &str,
-) -> AppResult<i64> {
-    let existing: Option<i64> = sqlx::query_scalar("SELECT id FROM artwork WHERE hash = ?")
-        .bind(hash)
-        .fetch_optional(&mut *conn)
-        .await?;
-    if let Some(id) = existing {
-        return Ok(id);
-    }
-
-    let now = now_millis();
-    let result =
-        sqlx::query("INSERT INTO artwork (hash, format, source, created_at) VALUES (?, ?, ?, ?)")
-            .bind(hash)
-            .bind(format)
-            .bind(source)
-            .bind(now)
-            .execute(&mut *conn)
-            .await?;
-    Ok(result.last_insert_rowid())
-}
-
-/// Split a raw artist string like `"Elior, DJ Garlik"` into individual
-/// names. Conservative: only splits on `", "` and `"; "` so that artist
-/// names containing `&`, `/`, or `feat.` (e.g. `"AC/DC"`, `"Simon &
-/// Garfunkel"`) stay intact.
-///
-/// Returns the trimmed, non-empty names in the order they appeared —
-/// the first entry is treated as the primary artist by the caller.
-pub(crate) fn split_artist_name(raw: &str) -> Vec<String> {
-    raw.split([',', ';'])
-        .map(|s| s.trim())
-        .filter(|s| !s.is_empty())
-        .map(|s| s.to_string())
-        .collect()
-}
-
-pub(crate) async fn upsert_artist(
-    conn: &mut sqlx::SqliteConnection,
-    raw_name: &str,
-) -> AppResult<Option<i64>> {
-    let name = raw_name.trim();
-    if name.is_empty() {
-        return Ok(None);
-    }
-    let canon = canonical_name(name);
-    if canon.is_empty() {
-        return Ok(None);
-    }
-
-    let existing: Option<i64> =
-        sqlx::query_scalar("SELECT id FROM artist WHERE canonical_name = ?")
-            .bind(&canon)
-            .fetch_optional(&mut *conn)
-            .await?;
-    if let Some(id) = existing {
-        return Ok(Some(id));
-    }
-
-    let result = sqlx::query("INSERT INTO artist (name, canonical_name) VALUES (?, ?)")
-        .bind(name)
-        .bind(&canon)
-        .execute(&mut *conn)
-        .await?;
-    Ok(Some(result.last_insert_rowid()))
-}
-
-/// Post-scan pass that promotes "tagless" same-title album rows into a
-/// single Various-Artists compilation. Catches the common case where a
-/// lofi / mood / cover-pack compilation (Soothing Breeze, Coffee Shop,
-/// etc.) ships without `aART` and without `TCMP` so each track lands
-/// in its own primary-artist-keyed album row.
-///
-/// Heuristic — conservative to avoid false positives on legit cases
-/// like "two different artists who self-titled":
-///   - same `canonical_title`
-///   - every row has `album_artist IS NULL AND is_compilation = 0`
-///     (the tag-driven path is the source of truth — never override)
-///   - at least 3 distinct `artist_id`s (so a featuring on one track
-///     of a normal album doesn't get promoted to a fake compilation)
-///
-/// On match: pick the lowest-id row as survivor, set
-/// `(artist_id = VariousArtists, album_artist = "Various Artists",
-/// is_compilation = 1)`, reparent every track of the sibling rows
-/// onto the survivor, and delete the siblings. Their artwork rows
-/// stay around (other albums may share them via hash dedup).
-async fn merge_implicit_compilations(pool: &SqlitePool) -> AppResult<()> {
-    let groups: Vec<(String,)> = sqlx::query_as(
-        r#"
-        SELECT canonical_title
-          FROM album
-         WHERE album_artist IS NULL
-           AND is_compilation = 0
-           AND artist_id IS NOT NULL
-         GROUP BY canonical_title
-        HAVING COUNT(DISTINCT artist_id) >= 3
-        "#,
-    )
-    .fetch_all(pool)
-    .await?;
-
-    if groups.is_empty() {
-        return Ok(());
-    }
-
-    let mut tx = pool.begin().await?;
-    let va_id = upsert_artist(&mut tx, VARIOUS_ARTISTS_LABEL).await?;
-    let Some(va_id) = va_id else {
-        // upsert_artist returned None — name canonicalised to empty,
-        // shouldn't happen with "Various Artists" but guard defensively.
-        tx.rollback().await?;
-        return Ok(());
-    };
-
-    for (canonical_title,) in groups {
-        let album_ids: Vec<i64> = sqlx::query_scalar(
-            "SELECT id FROM album
-              WHERE canonical_title = ?
-                AND album_artist IS NULL
-                AND is_compilation = 0
-              ORDER BY id ASC",
-        )
-        .bind(&canonical_title)
-        .fetch_all(&mut *tx)
-        .await?;
-
-        let Some((&survivor, siblings)) = album_ids.split_first() else {
-            continue;
-        };
-        if siblings.is_empty() {
-            continue;
-        }
-
-        sqlx::query(
-            "UPDATE album
-                SET artist_id    = ?,
-                    album_artist = ?,
-                    is_compilation = 1
-              WHERE id = ?",
-        )
-        .bind(va_id)
-        .bind(VARIOUS_ARTISTS_LABEL)
-        .bind(survivor)
-        .execute(&mut *tx)
-        .await?;
-
-        for sid in siblings {
-            sqlx::query("UPDATE track SET album_id = ? WHERE album_id = ?")
-                .bind(survivor)
-                .bind(sid)
-                .execute(&mut *tx)
-                .await?;
-            sqlx::query("DELETE FROM album WHERE id = ?")
-                .bind(sid)
-                .execute(&mut *tx)
-                .await?;
-        }
-
-        tracing::info!(
-            canonical_title = %canonical_title,
-            survivor,
-            merged = siblings.len(),
-            "auto-merged implicit compilation"
-        );
-    }
-
-    tx.commit().await?;
-    Ok(())
-}
-
-/// Sentinel album-artist row used when an album is tagged as a
-/// compilation but has no explicit Album Artist. Resolved to a real
-/// `artist` row on first encounter via [`upsert_artist`], then reused.
-const VARIOUS_ARTISTS_LABEL: &str = "Various Artists";
-
-/// Resolve the album-artist text to an `artist` row id, applying the
-/// scanner's grouping policy:
-///
-/// 1. Explicit Album Artist tag → upsert that name verbatim.
-/// 2. No tag but `is_compilation == true` → upsert the
-///    `"Various Artists"` sentinel so a TCMP-flagged record stays
-///    glued together regardless of per-track Artist diversity.
-/// 3. No tag and not a compilation → fall back to the first artist of
-///    the track (`track_primary_artist_id`), preserving the v1.0
-///    behaviour for files the user hasn't re-tagged yet.
-///
-/// Returns the chosen `artist.id` plus the display text we want to
-/// persist on `album.album_artist` (preserves the source casing).
-async fn resolve_album_artist(
-    conn: &mut sqlx::SqliteConnection,
-    album_artist: Option<&str>,
-    is_compilation: bool,
-    track_primary_artist_id: Option<i64>,
-) -> AppResult<(Option<i64>, Option<String>)> {
-    if let Some(name) = album_artist {
-        let name = name.trim();
-        if !name.is_empty() {
-            let id = upsert_artist(conn, name).await?;
-            return Ok((id, Some(name.to_string())));
-        }
-    }
-    if is_compilation {
-        let id = upsert_artist(conn, VARIOUS_ARTISTS_LABEL).await?;
-        return Ok((id, Some(VARIOUS_ARTISTS_LABEL.to_string())));
-    }
-    Ok((track_primary_artist_id, None))
-}
-
-pub(crate) async fn upsert_album(
-    conn: &mut sqlx::SqliteConnection,
-    title: &str,
-    album_artist_text: Option<&str>,
-    is_compilation: bool,
-    track_primary_artist_id: Option<i64>,
-    year: Option<i64>,
-) -> AppResult<Option<i64>> {
-    let title = title.trim();
-    if title.is_empty() {
-        return Ok(None);
-    }
-    let canon = canonical_name(title);
-    if canon.is_empty() {
-        return Ok(None);
-    }
-
-    let (artist_id, album_artist_display) = resolve_album_artist(
-        conn,
-        album_artist_text,
-        is_compilation,
-        track_primary_artist_id,
-    )
-    .await?;
-
-    // The `UNIQUE (canonical_title, artist_id)` constraint treats NULL as
-    // distinct in SQLite, so we dedup manually for the NULL-artist case.
-    let mut existing: Option<i64> = if let Some(aid) = artist_id {
-        sqlx::query_scalar("SELECT id FROM album WHERE canonical_title = ? AND artist_id = ?")
-            .bind(&canon)
-            .bind(aid)
-            .fetch_optional(&mut *conn)
-            .await?
-    } else {
-        sqlx::query_scalar("SELECT id FROM album WHERE canonical_title = ? AND artist_id IS NULL")
-            .bind(&canon)
-            .fetch_optional(&mut *conn)
-            .await?
-    };
-
-    // Re-use an existing compilation row for this title even when the
-    // incoming track has no Album Artist tag — without this, every
-    // rescan of a previously auto-merged compilation would re-fragment
-    // (the artist-specific SELECT above misses because the merged row
-    // has artist_id = "Various Artists"). Only applies when the
-    // incoming track has no explicit album_artist tag and isn't itself
-    // flagged as compilation; otherwise the explicit fields take
-    // precedence and the regular upsert path runs.
-    if existing.is_none() && album_artist_text.is_none() && !is_compilation {
-        existing = sqlx::query_scalar(
-            "SELECT id FROM album
-              WHERE canonical_title = ? AND is_compilation = 1
-              LIMIT 1",
-        )
-        .bind(&canon)
-        .fetch_optional(&mut *conn)
-        .await?;
-    }
-    if let Some(id) = existing {
-        // Backfill album_artist / is_compilation on the existing row
-        // ONLY when this scan brings new information. Re-scans of files
-        // without an Album Artist tag and without the compilation flag
-        // skip the UPDATE entirely — preserves the v1.0 perf profile
-        // for libraries the user hasn't re-tagged. The COALESCE / OR
-        // keeps the values "sticky": once a track declares an album
-        // artist or compilation, the row keeps it even if siblings
-        // drop the tags.
-        if album_artist_display.is_some() || is_compilation {
-            sqlx::query(
-                "UPDATE album
-                    SET album_artist   = COALESCE(album_artist, ?),
-                        is_compilation = CASE WHEN ? = 1 OR is_compilation = 1 THEN 1 ELSE 0 END
-                  WHERE id = ?",
-            )
-            .bind(album_artist_display.as_deref())
-            .bind(if is_compilation { 1_i64 } else { 0_i64 })
-            .bind(id)
-            .execute(&mut *conn)
-            .await?;
-        }
-        return Ok(Some(id));
-    }
-
-    let result = sqlx::query(
-        "INSERT INTO album (title, canonical_title, artist_id, year, album_artist, is_compilation)
-         VALUES (?, ?, ?, ?, ?, ?)",
-    )
-    .bind(title)
-    .bind(&canon)
-    .bind(artist_id)
-    .bind(year)
-    .bind(album_artist_display.as_deref())
-    .bind(if is_compilation { 1_i64 } else { 0_i64 })
-    .execute(&mut *conn)
-    .await?;
-    Ok(Some(result.last_insert_rowid()))
-}
-
-/// Walk every artist name parsed from `raw`, pair it with its `artist.id`
-/// from `artist_ids` (positionally aligned), and try to resolve a sidecar
-/// artist image from `track_path`. Idempotent — artists that already have
-/// an `artwork_id` are skipped by [`link_local_artist_image`].
-///
-/// Skips the "Various Artists" sentinel: a compilation folder never holds
-/// a meaningful artist photo and we'd just pin a random album cover to it.
-pub(crate) async fn maybe_link_artist_images(
-    conn: &mut sqlx::SqliteConnection,
-    artist_raw: Option<&str>,
-    artist_ids: &[i64],
-    track_path: &Path,
-    artwork_dir: &Path,
-) -> AppResult<()> {
-    let Some(raw) = artist_raw else {
-        return Ok(());
-    };
-    let names = split_artist_name(raw);
-    let va_canon = canonical_name(VARIOUS_ARTISTS_LABEL);
-    for (name, id) in names.iter().zip(artist_ids.iter()) {
-        let canon = canonical_name(name);
-        if canon.is_empty() || canon == va_canon {
-            continue;
-        }
-        // Cheap pre-check so we don't walk the FS when the artist already
-        // has artwork (Deezer fetch, manual upload, or earlier scan).
-        let has_artwork: Option<i64> =
-            sqlx::query_scalar("SELECT 1 FROM artist WHERE id = ? AND artwork_id IS NOT NULL")
-                .bind(id)
-                .fetch_optional(&mut *conn)
-                .await?;
-        if has_artwork.is_some() {
-            continue;
-        }
-        if let Some(cover) = extract_artist_image(track_path, &canon, artwork_dir) {
-            link_local_artist_image(&mut *conn, *id, &cover).await?;
-        }
-    }
-    Ok(())
-}
-
-/// Resolve a raw multi-artist string (e.g. `"A, B; C"`) to a vector of
-/// artist row IDs. The first entry becomes the track's primary artist.
-/// Empty / whitespace-only inputs yield an empty vector.
-pub(crate) async fn upsert_artist_list(
-    conn: &mut sqlx::SqliteConnection,
-    raw: &Option<String>,
-) -> AppResult<Vec<i64>> {
-    let Some(raw) = raw else {
-        return Ok(Vec::new());
-    };
-    let mut ids = Vec::new();
-    for name in split_artist_name(raw) {
-        if let Some(id) = upsert_artist(&mut *conn, &name).await? {
-            ids.push(id);
-        }
-    }
-    Ok(ids)
-}
-
-pub(crate) async fn upsert_genre(
-    conn: &mut sqlx::SqliteConnection,
-    raw_name: &str,
-) -> AppResult<Option<i64>> {
-    let name = raw_name.trim();
-    if name.is_empty() {
-        return Ok(None);
-    }
-    let canon = canonical_name(name);
-    if canon.is_empty() {
-        return Ok(None);
-    }
-
-    let existing: Option<i64> = sqlx::query_scalar("SELECT id FROM genre WHERE canonical_name = ?")
-        .bind(&canon)
-        .fetch_optional(&mut *conn)
-        .await?;
-    if let Some(id) = existing {
-        return Ok(Some(id));
-    }
-
-    let result = sqlx::query("INSERT INTO genre (name, canonical_name) VALUES (?, ?)")
-        .bind(name)
-        .bind(&canon)
-        .execute(&mut *conn)
-        .await?;
-    Ok(Some(result.last_insert_rowid()))
 }
 
 /// Walk an existing `library_folder` on disk, extract tags from every audio
@@ -1879,167 +1015,4 @@ pub async fn rescan_local_artist_images(
         "rescan_local_artist_images complete",
     );
     Ok(summary)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn write_bytes(path: &Path, bytes: &[u8]) {
-        fs::write(path, bytes).expect("write fixture");
-    }
-
-    /// Smallest valid 1x1 JPEG — enough to satisfy the non-empty check
-    /// and exercise the hash + write + spawn_thumbnail_job pipeline
-    /// without dragging the `image` crate into the unit test.
-    const TINY_JPEG: &[u8] = &[
-        0xFF, 0xD8, 0xFF, 0xE0, 0x00, 0x10, 0x4A, 0x46, 0x49, 0x46, 0x00, 0x01, 0xFF, 0xD9,
-    ];
-
-    #[test]
-    fn folder_cover_picks_priority_stem_over_alphabetical_first() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let folder = dir.path().join("album");
-        fs::create_dir_all(&folder).unwrap();
-
-        // `albumart` ranks below `cover` in FOLDER_COVER_STEMS; even though
-        // it sorts first alphabetically, the priority list must win.
-        write_bytes(&folder.join("albumart.jpg"), TINY_JPEG);
-        write_bytes(&folder.join("cover.png"), TINY_JPEG);
-
-        let track = folder.join("01.flac");
-        write_bytes(&track, b"not really audio");
-
-        let cover = extract_folder_cover(&track, &artwork_dir).expect("cover found");
-        assert_eq!(
-            cover.format, "png",
-            "cover.png should win over albumart.jpg"
-        );
-        assert_eq!(cover.source, "folder");
-    }
-
-    #[test]
-    fn folder_cover_normalises_jpeg_extension() {
-        let dir = tempfile::tempdir().unwrap();
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let folder = dir.path().join("album");
-        fs::create_dir_all(&folder).unwrap();
-
-        write_bytes(&folder.join("front.JPEG"), TINY_JPEG);
-        let track = folder.join("01.flac");
-        write_bytes(&track, b"x");
-
-        let cover = extract_folder_cover(&track, &artwork_dir).expect("cover found");
-        // `jpeg` must collapse to `jpg` so the artwork dir has one
-        // canonical extension per MIME.
-        assert_eq!(cover.format, "jpg");
-    }
-
-    #[test]
-    fn folder_cover_returns_none_when_no_match() {
-        let dir = tempfile::tempdir().unwrap();
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let folder = dir.path().join("album");
-        fs::create_dir_all(&folder).unwrap();
-
-        // Recognised extension but stem isn't in the priority list.
-        write_bytes(&folder.join("scan-of-booklet.jpg"), TINY_JPEG);
-        let track = folder.join("01.flac");
-        write_bytes(&track, b"x");
-
-        assert!(extract_folder_cover(&track, &artwork_dir).is_none());
-    }
-
-    #[test]
-    fn artist_image_finds_stem_in_parent_folder() {
-        // Layout: <root>/<Artist>/<Album>/<track>
-        let dir = tempfile::tempdir().unwrap();
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let artist_dir = dir.path().join("Daft Punk");
-        let album_dir = artist_dir.join("Discovery");
-        fs::create_dir_all(&album_dir).unwrap();
-
-        write_bytes(&artist_dir.join("artist.jpg"), TINY_JPEG);
-        let track = album_dir.join("01.flac");
-        write_bytes(&track, b"x");
-
-        let cover = extract_artist_image(&track, &canonical_name("Daft Punk"), &artwork_dir)
-            .expect("artist image found two levels up");
-        assert_eq!(cover.source, "folder");
-        assert_eq!(cover.format, "jpg");
-    }
-
-    #[test]
-    fn artist_image_matches_canonical_name_stem() {
-        // Layout: <root>/<Album>/<track> with `<Artist>.jpg` beside the album.
-        let dir = tempfile::tempdir().unwrap();
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let album_dir = dir.path().join("Discovery");
-        fs::create_dir_all(&album_dir).unwrap();
-
-        write_bytes(&album_dir.join("Daft Punk.png"), TINY_JPEG);
-        let track = album_dir.join("01.flac");
-        write_bytes(&track, b"x");
-
-        let cover = extract_artist_image(&track, &canonical_name("daft punk"), &artwork_dir)
-            .expect("canonical-name stem match");
-        assert_eq!(cover.format, "png");
-    }
-
-    #[test]
-    fn artist_image_ignores_unrelated_named_image() {
-        let dir = tempfile::tempdir().unwrap();
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let album_dir = dir.path().join("Discovery");
-        fs::create_dir_all(&album_dir).unwrap();
-
-        // `cover.jpg` is an album cover, not an artist photo.
-        write_bytes(&album_dir.join("cover.jpg"), TINY_JPEG);
-        let track = album_dir.join("01.flac");
-        write_bytes(&track, b"x");
-
-        assert!(
-            extract_artist_image(&track, &canonical_name("Daft Punk"), &artwork_dir).is_none(),
-            "should not pick up album cover as artist image",
-        );
-    }
-
-    #[test]
-    fn artist_image_returns_none_for_empty_canonical() {
-        let dir = tempfile::tempdir().unwrap();
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let folder = dir.path().join("album");
-        fs::create_dir_all(&folder).unwrap();
-        write_bytes(&folder.join("artist.jpg"), TINY_JPEG);
-        let track = folder.join("01.flac");
-        write_bytes(&track, b"x");
-
-        // Empty canonical → defensive bail-out so we don't match every dir.
-        assert!(extract_artist_image(&track, "", &artwork_dir).is_none());
-    }
-
-    #[test]
-    fn folder_cover_writes_hash_addressed_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let artwork_dir = dir.path().join("artwork");
-        fs::create_dir_all(&artwork_dir).unwrap();
-        let folder = dir.path().join("album");
-        fs::create_dir_all(&folder).unwrap();
-
-        write_bytes(&folder.join("cover.jpg"), TINY_JPEG);
-        let track = folder.join("01.flac");
-        write_bytes(&track, b"x");
-
-        let cover = extract_folder_cover(&track, &artwork_dir).expect("cover");
-        let on_disk = artwork_dir.join(format!("{}.{}", cover.hash, cover.format));
-        assert!(on_disk.exists(), "hash-addressed file must be written");
-    }
 }
